+++ packages/vscode-generator/src/templates/extension/web/src/declarations.d.ts
@@ -1,4 +1,0 @@
-declare module '*.scss' {
-  export const content: { [className: string]: string };
-  export default content;
-}